Amend config.xml and add app id for ionic.io build
Amend config.xml to include a description of the app,
the author name, email, and website. Due to problems
with local development build environment, I decided
to use a service provided by ionic.io that allows
the developer to build the apk in the cloud from the
ionic command line interface. The cloud service uses
a standard build environment and allows the developer
to download the built apk from the server.