Revise calculate to calculate total from shopping cart
Define calculate to:
(i) iterate over model array (items);
(ii) calculate the item total for each item, by multiplying
the quantity of that item in the cart by the price for that item
(iii) update the total which is synced with the model.

"Any changes to the view are immediately reflected in the model,
and any changes in the model are propagated to the view. The model
is the single-source-of-truth for the application state, greatly
simplifying the programming model for the developer. You can think
of the view as simply an instant projection of your
model."
(see Angular docs: https://docs.angularjs.org/guide/databinding)

Modify changePrice to recalculate
Modify decreaseQty to recalculate
Remove deductFromTotal as its responsibilities are shared by other
functions defined in the controller

Reorder ng-click calls to reflect new division of labour in
tab-items.html